+++ CHANGELOG.md
@@ -7,8 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - Net energy content model: `NetEnergyModel::GrossEnergyFraction`
     - Parameter `forage.gross_energy`
+    - Parameter `hft.digestion.k_maintenance`
     - Parameter `hft.digestion.me_coefficient`
-    - Parameter `hft.digestion.ne_coefficient`
 
 ### Removed
 - The old “default” net energy model `get_net_energy_content_default()`.